Put the items in Kroo's in the chests
Just like the tomb des file the items were being placed beneath the
chests.

I left the potions and wands that Kroo starts on alone so he can still
be occasionally extra nasty.

This is a minor nerf to the player losing several random chest
contents worth of loot, but dismal swamp is the best anyway.

I added a guaranteed potion and spellbook to the far right chests in
order to preserve a (minor) reason to actually go over there.