Add the Mortal Blade
Based on the Mortal Blade from Sekiro, with some inspiration from Nightblood from Warbreaker.

* lawful tsurugi, appears as a 'sakura-hilted tsurugi'
* deals 2x damage to all and 3x damage to mortal or undead (mortal_race or is_undead), and the base damage has the energy mask added to it
** this only applies when it is wielded by the player and the player alone. for anybody else or when thrown/offhanded/etc., it will stay sheathed - treating it as a large club for damage purposes and skipping any artifact bonus damage
** this is shown properly in invent.c when wielded/not wielded. monster AI should prioritize this as a shitty large club rather than the cool version. under the hood, uses AD_DARK with custom damage handling elsewhere because it was the closest thing, but it's never shown as such to the player
* entirely bypasses most lifesaving sources and prevents slain monsters from reviving. for consumed items, this does not consume them and they will (probably) drop as normal
** exceptions are the yog twin lifesaving & the suzerain puppet lifesaving, since those warp them away, and to prevent weirdo behavior it just... lets them vanish. yog and the suzerain are like, getting replacements or something.
** riders still revive as normal. this affects trolls, zombie reviving, fungus reviving, but not stuff growing on them etc.
* will kill you when you wield it. this is unpreventable - any ways to avoid this are oversights. specifically, it calls done(DIED), so it will kill whoever does it without a save, identically to something like the Silence Glaive. notably, this can be bypassed via lifesaving from any source, which will leave the weapon wielded in your hands. if you unwield it... you will lose the bonus, and will kill yourself again on a rewield.

changes made in the process -
adds a new field to monsters, mlaidtorest. this is used for ARTA_LAIDTOREST (only for the mortal blade, but feels like the kind of thing that could go on another artifact so it's an ARTA flag). it is responsible for the lifesaving/revive bypass. it works by being set in the artifact extra damage/effect code on hit, and is cleared at the end of the attack's resolution (post-corpse creation, notably). this means anything that dies with it in effect can't revive etc., but that should only be possible on the attack that set it.
fungi/migos no longer revive if cancelled. cancellation kills all the extra cells or something idk it seems appropriate, like come on, nobody cancels things anyway. let me have this
dark damage, for consistency with AD_DARK monster attacks, is now doubled for artifact damage like holy damage. xhity.c AD_DARK attacks are 1x damage vs. non-dark immune (unliving, primordial) and then doubled to 2x vs. dark-vuln (intelligent non-GOO living things). artifact damage now does the same, which only really affects the Dark Claymore & Crucifix of the Mad King (1d6 and 1d10 dark damage respectively... they're barely noticing it, come on). This is specifically special-cased for the Mortal Blade so it's what was said above, since I wanted a weird mish-mash of the AD_DARK criteria + some more so w/e. not like the player knows it's AD_DARK overloaded secretly

this is NOT at all playtested. as of writing this i don't plan on making it a PR until i do at least one run (probably sflame human samurai..) with it, but I think it's safe to say this falls in the category of "probably not a good weapon for anybody really". it does have some interesting synergy with sflame, since they can get easy curse res for this 2handed weapon + lifesaving on their hat re-applied, but that might start getting real expensive real quick... maybe i'll come back with an "actually, fix sflame prices for their major oprops" isntead or something.

very savebreaking obviously lol

might continue this commit later with an invoke for empowered mortal draw or something. not sure.

---- cherrypicked from a 3.22.1 branch for merge