Convict revision
Ignoring the quest is a bad idea now.

Technically, this is Hellraiser inspired.

Also, technically unusually spoilable, but the timeline here is pretty tight so it's PROBABLY best not to have to find out about it the hard way.

The "demon" summoning ritual in the basement progresses over the first part of the game, completing on turn 21,000. The effects of the ritual then spread upward, engulfing the home level on turn 30,000 and beginning to affect the dungeons of doom on turn 42,000. Importantly, if the quest isn't accepted by turn 30,000 the iron ball of levitation is lost, though the PC will now be able to enter the quest proper without permision.

The Nemesis is psychic now and gets an extra spell choice every 3000 turns until the ritual completes.
-(psi, doubt, san, barf, pain, and spacewarp, in that order)

Starting turn 33000 the quest will suffer extra spawns (the ritual effects have spread into the city)

Adjust drohea robe to be a generic gift for convicts.

Adds the voice-in-screams, loosely based on the pathfinder velstrac demagogue of the same title.
-30th level, two 2d25 physical touch attacks (ouch)
-Adds beartrap, arrow rain (flaying scalpels), and punish to the Warden's list
-Beartrap spell traps the target in a beartrap, or creates a chain golem if they're already in a trap (works on monsters and the PC)

Adds chain devils
-Pretty standard, has an unerving gaze attack
-Gaze damages morale and may force a save-vs-sanloss

Adds Cuboid and Rhombohedroid monsters
-Malkuth and Hod Keter variants
-Immobile
-Has the revevant melee attacks and unnerving beam attacks
-Beam attacks lower morale and force a save vs. sanloss

Adds a flayed template
-+9 AC and DR, always hostile, follows between levels, demon (MA_, hates holy, hates silver, regen), +6 move
-Psionic casting: Open wounds, plus doubt, san, barf, and pain bolts

Adds a nuncio faction (nuncio: a papal messenger; what dnh calls cenobites/kytons/velstracs)

Buffs Luckblade
-Imposes "disadvantage" on monster melee and spellcasting rolls (roll twice and take worst)
-Makes monsters fumbling (1/20 chance of each effect)
--They sometimes slip and lose turns like the PC
--Plus a variety of slapstick accidents
---Drop wielded weapons
---Hit adjacent monsters in melee
---Drop items instead of using them (including AoLS)
---Ranged weapons may slip/misfire
-+1 to untrapping
-Always tames prisoners
-Give +Luck to pet xp_threshold
-Adds graze etrait (which convicts typically can't use due to low skill)
-Wielding it effectively allows you to roll twice when casting spells (by adjusting the failure odds)

Improves the "disguse" code so that more things can be used to hide from the minetown guards
-Masks, sunglasses, and r'lyehian faceplates, opaque face-covering helms, opaque faceless cloaks

Carrying the spoon of liberation boosts, tp away, stone to flesh, and dig

Adds some missing unalive tags on mechanical monsters

Allow convicts to make the brands into heavy balls

Uncap heavy ball damage. It's probably fine.

Convert a couple generic prisoners in the convict quest to throne-room prisoners.