Fully adopt a "must have lifted the veil" req for all research
Add to x_ok()s and messages to research self-exam.

I was mislead by the missing messages to believe that it should be possible, but it wasn't intended to be.

Pre-quest upgrades are meant to be handled by smithing. You can easily get +2 metal armor in every slot, plus a better-than-+2 weapon. This should be fine.

A future idea is to have dissected corpses drop upgrade materials or something, so that there *is* a slight sac-fest equivalent for UH.