Undead hunters improve at least 3 skill caps to expert over the course of the game
One existing cap of skilled is chosen and improved to expert.

If there is an uh weapon of that skill with a paired weapon of another skill, the second skill may be improved as well
-"may": if there is also an uh weapon of that skill with NO paired skill, the second may not be improved. If there are two paired weapons with different skills, only one will be improved.

These improvements are handed out at the same time the smithing skill is improved ("The knowledge from the crystal sinks into your subconscious.").
-1 (pair) of skills is chosen when improving smithing from basic to skilled with a twin columnar crystal.
-2 more (pairs) of skills are chose when improving smithing from skilled to expert with a chunk or mass.
--If the first improvement is skipped (via enchant weapon/armor scrolls) all three improvents are handed out when improving smithing from basic->expert.