Fix altar display in mapseen
Previously, #overview would show the altar's god as being to your god
if it's of your alignment (regardless of the altar's actual god), and
doesn't show anything about the altar's god/alignment otherwise.  This
commit makes it always show the altar's correct god if all altars on
the level are to the same god.

mapseen-related stuff is moved from dungeon.h to mapseen.h, to avoid a
circular makedefs -> dungeon.h -> gnames.h -> makedefs dependency.
gnames.h needs to be included for MAX_GOD.

# Conflicts:
#	include/tradstdc.h