Slightly less insane noattack handling
noattack is now declared const

It is generally only ever used to initialize attack buffers

A dedicated noattack entry was added to the spirit attacks array.

The spirit attack function's return value is likewise used to initialize an attack buffer.

The original u/mhitu/m code assumed that attack structs generally shouldn't be modified.

The current xhity code assumes that the current attack pointer points to a buffer that can be freely modified.

This was, shall we say, a bad combination.