Use a static runmove counter to re-draw the screen every 7 turns
As intended by the original author. The use of moves here caused the screen to redraw multiple times a turn if the PC got to take multiple steps, eg from boots of speed.

This effect is only somewhat noticeable in vanilla but extremely noticiable in dnh due to 7L boots etc.