Restock dungeon levels "while" the PC is gone
(Actually spawns monsters when they revist the level)

Records the last turn on which the random monster check happened on every dungeon floor.
-The Ana quest is on its own timeline.

Refactors the random monster spawn code and random monster frequency code into their own functions.

Has a 66% chance to spawn a random monster per frequency turns.
-Levels with below-average spawn rates (the pacified fem hd nob quest) are exempt.

Caps out at 10 spawns+pre-existing hostile monsters on the level.

All in all this makes for a very noticeable restock rate. Hit and run tactics will be less effective now that both levels can respawn enemies, and traversing "cleared" levels to return to the stash etc is more combat-heavy

It's plausibly the case this is too fast, and/or that there should be a "clearing" mechanic.