Space-for-time trade in menu color code
Use a hashmap linking strings to colors and attributes so that we can spend less time evaluating regexs.

Also record the last turn on which a coloring was used, and every 1000 turns free all entries that haven't been used for more than 1000 turns.

This uses https://github.com/sheredom/hashmap.h and includes it in dnhethack's sources. As it is released into the public domain under unlicense, I believe this causes no problems.