Fix hashmap rake
It's not initialized for the first few menus so it crashes when you try to start a new game (guess what I was doing during initial testing >_<).

Could fix it by tracing things back and initializing the hashmap sooner, or could fix it by ensuring that we don't crash if we don't have a hashmap yet.

I picked the latter.