Add a dissection kit, place one in the uh quest
Applying this kit while to a corpse dissects it and advances all active research tracks.

Applying this kit over a philosophy altar refills it.

Late game as an undead hunter you may need to dissect an excessive number of monsters. The kit exists to simplify that, since you won't need to altar camp so much.

One is spawned in the possetion of a new "monster" (crimson moon-lotus) that mostly exists to hold the kit behind the veil for you.

Also touches up research
-Dissecting c without gloves is fatal now
-College gets extra research credit for the first of every monster type dissected and reanimated.