The college can unlock more undead hunter weapon skills.
If you have a build you really want to use, and you really want expert skill, this is a way to do it.

It doesn't require exponentially increasing research to do it, so you can fairly quickly unlock all undead hunter skills to expert if you REALLY want. The dissection kit, reanimation tool, and dungeon restocking really help with this.