Implement silverflame+blood smithing
Previously, the PC could move around Goat and Yog upgrades, but not silver flame upgrades.

All silver flame weapon upgrades are extracted. Armor upgrades are left behind (and may require a new base-level silver flame to reactivate, but those are basically free)

Because not all silver flame crystals are equal, use a menu to allow the player to choose which one to use.
-They can name them to disambiguate if they want.
-In practice, I doubt the PC will ever have more than one such crystal in inventory, but it'll work smoothly if they DO have more than one.

This also limits the ability to extract upgrades to gear that the PC could apply upgrades to.
-Prevents the player accidently destroying a piece of upgraded armor or something.