Make a vermiurge fossil in the rocks of the first level of the law quest.
Some people want one guaranteed for scorpion carapace reasons, but I don't want to put one on any level in particular. So, put a dead one in the random junk afterlife.

Tweaks make law fossil so that the fossil always gets spawned, even if there isn't a good spot for it.