Bugfix: Make rage walker rage quickly fade away, and up the insight threshold required.
Rage follows the PC between levels, so it can lead to rage walkers spawning in eg gatetown.

It's also rough for Binders and Madmen, who start with lifted veils

The net effect of these changes will be to ALMOST remove rage-walkers from the game, but oh well. The correct fix requires a save compatability break.