Mortal blade fixups - yesno prompt, can't gift, etc.
It absolutely needs a yesno prompt instead of just yn for drawing it - don't kill yourself accidentally. Similarly, just prompt if you try to unwield it.

It should probably never be gifted, because it's entirely unusable for most builds, it's too opt-in.

Also minor changes - if it's going to be so exorbitantly expensive, it gets to be phasing + more accurate. I don't think its a stretch power-wise, but maybe it is, dunno. You're still very 'married' to it.

Lastly, you can read it now (with insight broken, because fuck it) for the true name of the sword, and it shows as 'drawn' in inventory when drawn (alongside 'sheathed'). That might be unnecessary, but I feel like the visual clarity might help you not autopilot unwield it.